Fix issue for install gppkg in new gpdb
In this commit, the version info of postgis gppkg is fixed about orca version
issue. Moreover, the makefiles are also fixed and allow gppkg to copy the *.so
and *.sql file into gppkg. After this commit, the postgis gppkg can be
successfully installed in a new gpdb without need the source code of postgis.

Authors: Kuien Liu and Haozhou Wang

quick fix to build gppkg faster and add orca to gppkg name